MEApp: remove unused code: props and emits
This currently gets its data injected, not via props, and also, it emits
nothing. Maybe this should be reworked. Reloading could/should move to
the store, or reloading could be handled by the surrounding application.
But as of now, props and emit are dead code and thus better deleted. Can
be readded again if needed later.