Improve preview of datavalue
It is confusing when a snaktype is set to "no value" but the rendered
preview still shows what it was before.

The alternative would have been to just hide the preview section when a
snaktype other than "customvalue" is selected. But that would result in
a conceptual "disconnect" between what is in the store and what is
displayed to the user.

Also, this clears the qualifier heading because the float from the
geo-coordinate preview causes it to move to unexpected places otherwise.

The unit `lh` is only supported on quite modern browsers, but there it
should reliably reduce UI movements for many datatypes, and for older
browsers there should things not get worse compared to the status quo.