Drop the selector for the absence MediaWiki theme class
Dark mode is already enabled on www.wikidata.org, but so broken that it is probably
basically not used by anyone. However, that means that we can't the
presence of the class as an indicator for the overall desire of the user
for light/dark mode. Thus this change now follows only the device
setting and not the mediawiki setting.

This will mean that it is less future-proof and will need explicit work
on dropping this code once the dark mode on www.wikidata.org becomes
usable by default.